Fixing bad import
