+++ src/deephaven_ib/_tws/order_id_queue.py
@@ -2,16 +2,33 @@
 
 from threading import Event, Thread
 from time import sleep
+from enum import Enum
 from typing import List, Callable, TYPE_CHECKING
 
 from .._internal.threading import LoggingLock
 from .._internal.trace import trace_all_threads_str
-from .. import OrderIdStrategy
 
 # Type hints on IbTwsClient cause a circular dependency.
 # This conditional import plus a string-based annotation avoids the problem.
 if TYPE_CHECKING:
     from .tws_client import IbTwsClient
+
+
+class OrderIdStrategy(Enum):
+    """Strategy used to obtain order IDs."""
+
+    def __new__(cls, retry:bool, tws_request:bool):
+        obj = bytes.__new__(cls)
+        obj.retry = retry
+        obj.tws_request = tws_request
+        return obj
+
+    INCREMENT = (False, False)
+    """Use the initial next order ID and increment the value upon every call.  This is fast, but it may fail for multiple sessions."""
+    BASIC = (False, True)
+    """Request a new order IDs from TWS every time one is needed."""
+    RETRY = (True, True)
+    """Request a new order IDs from TWS every time one is needed.  Retry if TWS does not respond quickly.  TWS seems to have a bug where it does not always respond."""
 
 
 class OrderIdRequest:
